Add files via upload
the base file for this project. for linux and unix users, you can store this command in a folder and add this folder to $PATH. instruction to add a folder to $PATH is here : 
Open your ~/.bashrc file and at the very end, add this: export PATH=/your/dir/path:$PATH  
Now,replace the /your/dir/path string with the original path of your directory where lknum is stored. save changes to the file and now you can run the command by simply typing lknum /path/to/your/file and you will get the number of links as the output.
NOTE: this program may not work on all systems, be sure to use /bin/bash and /bin/bash ony else the script will probably mess up